+++ CHANGES.md
@@ -1,6 +1,12 @@
 Libplanet changelog
 ===================
 
+Version 1.4.1
+-------------
+
+To be released.
+
+
 Version 1.4.0
 -------------